+++ CHANGES.md
@@ -10,6 +10,11 @@ To be released.
 
 ### Backward-incompatible API changes
 
+ -  Removed `IActionContext.Random` property.  Use `IActionContext.GetRandom()`
+    instead.  [[#3437]]
+ -  Added `IActionContext.RandomSeed` property.  [[#3437]]
+ -  Added `IActionContext.GetRandom()` method.  [[#3437]]
+
 ### Backward-incompatible network protocol changes
 
 ### Backward-incompatible storage format changes
@@ -23,6 +28,8 @@ To be released.
 ### Dependencies
 
 ### CLI tools
+
+[#3437]: https://github.com/planetarium/libplanet/pull/3437
 
 
 Version 3.4.0